Move cryspy.py to easydiffraction/calculators/cryspy/calculator.py
